requirements.txt
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.0
pytest>=7.0.0
python-dotenv>=0.19.0